@@ -12,8 +12,24 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1"
+    "@emotion/react": "^11.11.1",
+    "@emotion/styled": "^11.11.0",
+    "@material-ui/core": "^4.12.4",
+    "@mui/icons-material": "^5.13.7",
+    "@mui/material": "^5.13.7",
+    "axios": "^1.4.0",
+    "firebase": "^10.0.0",
+    "gapi-script": "^1.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-google-login": "^5.2.2",
+    "react-helmet": "^6.1.0",
+    "react-modal": "^3.16.1",
+    "react-router": "^6.13.0",
+    "react-router-dom": "^6.13.0",
+    "react-share": "^4.4.1",
+    "reactjs-popup": "^2.0.5",
+    "text-encoding": "^0.7.0"
   },
   "jest": {
     "moduleNameMapper": {
@@ -39,7 +55,10 @@
     "css-loader": "^4.3.0",
     "file-loader": "6.2.0",
     "html-webpack-plugin": "^4.5.0",
-    "style-loader": "^1.2.1",
+    "node-sass": "^9.0.0",
+    "sass": "^1.63.6",
+    "sass-loader": "^10.4.1",
+    "style-loader": "^1.3.0",
     "webpack": "^4.44.2",
     "webpack-cli": "^3.3.12",
     "webpack-dev-server": "^3.11.0"
