@@ -1,7 +1,12 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import ReactDOM from "react-dom/client";
 import App from "./components/App";
+import { BrowserRouter } from "react-router-dom";
 
 
-
-ReactDOM.render(<App />, document.getElementById("root"));
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(  
+    <BrowserRouter>
+      <App />
+    </BrowserRouter>
+);
